Reorganize project to fit babel-cli building workflow
