+++ package.json
@@ -2,10 +2,11 @@
   "name": "leanseed",
   "version": "0.2.0",
   "description": "a leanCloud seed project based on Express and Vue.js",
-  "main": "app.js",
+  "main": "build/index.js",
   "scripts": {
-    "start": "node app.js",
+    "start": "node build/index.js",
     "debug": "DEBUG=leanseed,leanseed:* node app.js",
+    "build-server": "node_modules/.bin/babel server -d build",
     "build": "NODE_ENV=production node_modules/.bin/webpack -p",
     "dev": "node_modules/.bin/webpack-dev-server --inline --hot --port 8080",
     "init-hotel": "node_modules/.bin/webpack-dev-server --inline --hot --port $PORT --host 0.0.0.0",
@@ -29,9 +30,10 @@
     "express": "4.13.1",
     "leanengine": "^1.0.0-beta",
     "body-parser": "^1.13.2",
-    "cookie-parser": "^1.3.5",
+    "cookie-parser": "^1.3.5"
   },
   "devDependencies": {
+    "babel-cli": "^6.0.0",
     "babel-core": "^6.4.0",
     "babel-register": "^6.4.3",
     "babel-preset-es2015": "^6.3.13",
